add a small c_api test
